chore: change the Flask app structure per their 'large applications' manual
