+++ src/annotation/.env.example
@@ -3,9 +3,6 @@ FLASK_APP=app.py
 GOOGLE_CLIENT_ID=1-2-3
 GOOGLE_CLIENT_SECRET=4-5-6
 
-# Uncomment to have code auto-reloaded but slower
-# FLASK_ENV=development
-
 # Uncomment it for local setups so it doesn't require HTTPS anymore
 # warning: any value counts as True, including 0, false, False and so on - so never set this variable on prod at all
 # OAUTHLIB_INSECURE_TRANSPORT=1
