+++ src/annotation/annotations_app/config.py
@@ -1,0 +1,1 @@
+# Ideally all env variables usage should be here
