+++ src/annotation/annotations_app/static/annotations.js
