+++ src/annotation/annotations_app/static/img/icon_box.png
