+++ src/annotation/annotations_app/static/img/icon_radial.png
