+++ src/annotation/annotations_app/static/init.js
